requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.4.6
MarkupSafe==1.0
virtualenv==16.7.2
Werkzeug==0.14.1
